+++ app/src/main/java/com/martiserramolina/lifeplan/viewmodels/utils/interfaces/CapableOfFetchingItemsI.kt
@@ -3,9 +3,7 @@ package com.martiserramolina.lifeplan.viewmodels.utils.interfaces
 import androidx.lifecycle.MutableLiveData
 
 interface CapableOfFetchingItemsI<T> {
-    companion object {
-        private const val NUM_ITEMS_TO_FETCH = 10
-    }
+    companion object { private const val NUM_ITEMS_TO_FETCH = 10 }
     val itemsFetched: MutableLiveData<MutableList<T>>
-    fun fetchItemsIfNotFetched(position: Int, numItems: Int = NUM_ITEMS_TO_FETCH): Boolean
+    fun fetchItemsIfNotFetched(position: Long, numItems: Int = NUM_ITEMS_TO_FETCH): Boolean
 }
